Initial working prototype for the milling post-process

+++ src/libslic3r/CustomGCode.cpp
@@ -56,13 +56,13 @@ extern void check_mode_for_custom_gcode_per_print_z(Info& info)
 
 // Return pairs of <print_z, 1-based extruder ID> sorted by increasing print_z from custom_gcode_per_print_z.
 // print_z corresponds to the first layer printed with the new extruder.
-std::vector<std::pair<double, unsigned int>> custom_tool_changes(const Info& custom_gcode_per_print_z, size_t num_extruders)
+std::vector<std::pair<double, uint16_t>> custom_tool_changes(const Info& custom_gcode_per_print_z, size_t num_extruders)
 {
-    std::vector<std::pair<double, unsigned int>> custom_tool_changes;
+    std::vector<std::pair<double, uint16_t>> custom_tool_changes;
     for (const Item& custom_gcode : custom_gcode_per_print_z.gcodes)
         if (custom_gcode.gcode == ToolChangeCode) {
             // If extruder count in PrinterSettings was changed, use default (0) extruder for extruders, more than num_extruders
-            custom_tool_changes.emplace_back(custom_gcode.print_z, static_cast<unsigned int>(custom_gcode.extruder > num_extruders ? 1 : custom_gcode.extruder));
+            custom_tool_changes.emplace_back(custom_gcode.print_z, static_cast<uint16_t>(custom_gcode.extruder > num_extruders ? 1 : custom_gcode.extruder));
         }
     return custom_tool_changes;
 }
