+++ src/libslic3r/CustomGCode.hpp
@@ -81,7 +81,7 @@ extern void check_mode_for_custom_gcode_per_print_z(Info& info);
 
 // Return pairs of <print_z, 1-based extruder ID> sorted by increasing print_z from custom_gcode_per_print_z.
 // print_z corresponds to the first layer printed with the new extruder.
-std::vector<std::pair<double, unsigned int>> custom_tool_changes(const Info& custom_gcode_per_print_z, size_t num_extruders);
+std::vector<std::pair<double, uint16_t>> custom_tool_changes(const Info& custom_gcode_per_print_z, size_t num_extruders);
 
 } // namespace CustomGCode
