+++ src/libslic3r/Slicing.cpp
@@ -62,7 +62,7 @@ SlicingParameters SlicingParameters::create_from_config(
 	const PrintConfig 		&print_config, 
 	const PrintObjectConfig &object_config,
 	coordf_t				 object_height,
-	const std::vector<unsigned int> &object_extruders)
+	const std::vector<uint16_t> &object_extruders)
 {
     //first layer height is got from the first_layer_height setting unless the value was garbage.
     // if the first_layer_height setting depends of the nozzle width, use the first one.
